requirements.txt
# Core Dependencies for BDU Chatbot RAG System
# Python 3.10+ required

# Web Framework
streamlit==1.51.0
extra_streamlit_components==0.1.81

# LLM & AI
groq==0.36.0
sentence-transformers==5.1.2
torch==2.9.1
transformers==4.57.1
huggingface-hub==0.36.0
scipy==1.15.3
scikit-learn==1.7.2
Pillow==12.0.0
typing-extensions==4.15.0

# Vector Database
qdrant-client==1.9.0
grpcio==1.76.0
protobuf==5.29.5

# Text Processing
langchain-text-splitters==1.0.0
python-dotenv==1.2.1

# Document Processing
pymupdf==1.26.6
python-docx==1.2.0
beautifulsoup4==4.14.2

# Data Processing
pandas==2.3.3
numpy==2.2.6
openpyxl==3.1.5

# Google API (for web search fallback)
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
httplib2==0.31.0

# Utilities
requests==2.32.5
tqdm==4.67.1
plotly==6.5.0
